requirements.txt
influxdb-client>=1.36.0
pytz>=2023.3
python-dotenv>=1.0.0
